server/requirements.txt
flask
gunicorn
flask_cors
datetime
werkzeug.security
app
qrcode
uuid
functools
pandas
bson.objectid
io
math
